Clarifies how to use a configuration file.
The original text referred to "schema.conf", which isn't actually found
anywhere else in the docs or even in t/ (where it's schema.pl).  I had
to post a question on StackOverflow to figure this out, even though the
answer is already in the docs, if not obvious from the original wording.